+++ pkgs/maa-pipeline-manager/CHANGELOG.md
@@ -2,3 +2,4 @@
 
 Fix tasks' pipeline_override missing
 Fix default of input missing
+Interface parser support group